server/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
common-lib==0.1.0
Django==4.0.3
Flask==2.1.1
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
idna==3.3
image==1.5.33
itsdangerous==2.1.2
Jinja2==3.1.1
jsonpickle==2.1.0
MarkupSafe==2.1.1
mysqlclient==2.1.0
numpy==1.22.3
opencv-contrib-python==4.5.5.64
Pillow==9.1.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
